Make compatble with gosdk head. (#729)
* fix(zcncore): set nonce on submitting

* merge

* merge

* revert config changes

+++ code/go/0chain.net/core/transaction/entity.go
@@ -153,7 +153,7 @@ func NewTransactionEntity() (*Transaction, error) {
 
 func (t *Transaction) ExecuteSmartContract(address, methodName string, input interface{}, val int64) error {
 	t.wg.Add(1)
-	err := t.zcntxn.ExecuteSmartContract(address, methodName, input, val)
+	err := t.zcntxn.ExecuteSmartContract(address, methodName, input, uint64(val))
 	if err != nil {
 		t.wg.Done()
 		return err
